requirements.txt
torch
numpy
pillow
tqdm